+++ src/Credfeto.Notification.Bot.Database/Twitch/DataManagers/TwitchStreamerDataManager.cs
@@ -19,10 +19,9 @@ public sealed class TwitchStreamerDataManager : ITwitchStreamerDataManager
         this._twitchUserBuilder = twitchUserBuilder ?? throw new ArgumentNullException(nameof(twitchUserBuilder));
     }
 
-    public Task AddStreamerAsync(Streamer streamerName, string streamerId, DateTime startedStreaming)
+    public Task AddStreamerAsync(Streamer streamerName, string streamerId, DateTimeOffset startedStreaming)
     {
-        return this._database.ExecuteAsync(storedProcedure: "twitch.streamer_insert",
-                                           new { username_ = streamerName.ToString(), id_ = streamerId, date_created_ = startedStreaming });
+        return this._database.ExecuteAsync(storedProcedure: "twitch.streamer_insert", new { username_ = streamerName.ToString(), id_ = streamerId, date_created_ = startedStreaming });
     }
 
     public Task<TwitchUser?> GetByUserNameAsync(Streamer userName)