+++ src/Credfeto.Notification.Bot.Shared/ICurrentTimeSource.cs
@@ -1,8 +1,0 @@
-using System;
-
-namespace Credfeto.Notification.Bot.Shared;
-
-public interface ICurrentTimeSource
-{
-    DateTime UtcNow();
-}