+++ src/Credfeto.Notification.Bot.Twitch.Data.Interfaces/ITwitchStreamerDataManager.cs
@@ -6,7 +6,7 @@ namespace Credfeto.Notification.Bot.Twitch.Data.Interfaces;
 
 public interface ITwitchStreamerDataManager
 {
-    Task AddStreamerAsync(Streamer streamerName, string streamerId, DateTime startedStreaming);
+    Task AddStreamerAsync(Streamer streamerName, string streamerId, DateTimeOffset startedStreaming);
 
     Task<TwitchUser?> GetByUserNameAsync(Streamer userName);