+++ src/Credfeto.Notification.Bot.Twitch.Tests/DependencyInjectionTests.cs
@@ -1,4 +1,5 @@
 using System;
+using Credfeto.Date.Interfaces;
 using Credfeto.Notification.Bot.Mocks;
 using Credfeto.Notification.Bot.Shared;
 using Credfeto.Notification.Bot.Twitch.Actions;